Ready to fill in implementations for RecordFillable.

+++ include/awkward/fillable/TupleFillable.h
@@ -14,7 +14,12 @@
 namespace awkward {
   class TupleFillable: public Fillable {
   public:
-    TupleFillable(const FillableOptions& options, const std::vector<std::shared_ptr<Fillable>>& contents, int64_t length, bool begun, size_t nextindex): options_(options), contents_(contents), length_(length), begun_(begun), nextindex_(nextindex) { }
+    TupleFillable(const FillableOptions& options, const std::vector<std::shared_ptr<Fillable>>& contents, int64_t length, bool begun, size_t nextindex):
+        options_(options),
+        contents_(contents),
+        length_(length),
+        begun_(begun),
+        nextindex_(nextindex) { }
 
     static TupleFillable* fromempty(const FillableOptions& options) {
       return new TupleFillable(options, std::vector<std::shared_ptr<Fillable>>(), -1, false, -1);
@@ -47,7 +52,7 @@ namespace awkward {
     std::vector<std::shared_ptr<Fillable>> contents_;
     int64_t length_;
     bool begun_;
-    size_t nextindex_;
+    int64_t nextindex_;
 
     void maybeupdate(int64_t i, Fillable* tmp);
   };
